FIX front tests

+++ daikoku/app/domain/entities.scala
@@ -157,7 +157,7 @@ case class Evolution(
 case class ReportsInfo(
                         id: DatastoreId,
                         activated: Boolean,
-                        date: Option[Double] = None,
+                        date: Option[Long] = None,
   ) extends CanJson[ReportsInfo] {
   override def asJson: JsValue = json.ReportsInfoFormat.writes(this)
 }
